incentivised county boundaries (issue #1, PR #9)
Now quantifying the scale of county boundary breaches, rather than only considering if a boundary is breached or not.

This is implemented by (implicitly) assigning a "primary" county to each constituency, and any EDs in a constituency not in the constituency's primary county increase the county boundary Hamiltonian $H_B$. This results in an incentive for constituencies to reduce breaches, addressing the main concern of the basic county boundary implementation (#4).