removing contiguity Hamiltonian
Removing all references to the contiguity Hamiltonian. This method of updating sites (i.e. only proposing from neighbours and not properly caring about contiguity) is the same as what was originally implemented at the hackathon, only now it's in C++ rather than Python, county boundaries and multi-seat constituencies are being considered, and the coupling constants are properly normalised.